editor/libeditor/tests/test_bug487524.html

<!DOCTYPE HTML>
<html><head>
<title>Test for bug 487524</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/EventUtils.js"></script>

<script class="testbody" type="application/javascript">
function runTest() {
  function setupIframe(e, html, focus_id) {
    var doc = e.contentDocument;
    doc.body.innerHTML = html;
    doc.designMode = "on";
    e = doc.getElementById(focus_id);
    doc.defaultView.focus();
    if (e) e.focus();
    return e;
  }

  var i1 = document.getElementById("i1");
  var li1 = setupIframe(i1, '<ul><li id="li1">one</li><li>two</li><ul><li>a</li></ul></ul>', "li1");
  var doc = li1.ownerDocument;

  var selection = doc.defaultView.getSelection();
  selection.removeAllRanges();

  var range = doc.createRange();
  range.setStart(li1, 0);
  range.setEnd(li1.nextSibling, 0);
  selection.addRange(range);

  sendKey("delete");
  is(doc.body.innerHTML, "<ul><li>two</li><ul><li>a</li></ul></ul>", "delete 1st LI");

  var li2 = setupIframe(i1, '<ul><li id="li2">two</li><ul><li>a</li></ul></ul>', "li2");
  selection = doc.defaultView.getSelection();
  selection.removeAllRanges();

  range = doc.createRange();
  range.setStart(li2, 0);
  range.setEnd(li2.nextSibling.firstChild, 0);
  selection.addRange(range);

  sendKey("delete");
  is(doc.body.innerHTML, "<ul><ul><li>a</li></ul></ul>", "delete 2nd LI");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);
</script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=487524">Mozilla Bug 487524</a>
<p id="display"></p>

<pre id="test">
</pre>

<iframe id="i1" width="200" height="100" src="about:blank"></iframe><br>

</body>
</html>
